apps/asml_showcase_app/requirements.txt
streamlit>=1.38.0
databricks-sdk[openai]>=0.81.0
psycopg[binary]>=3.2.0
mlflow[databricks]>=3.9.0
